+++ app/class/common.rb
@@ -1,0 +1,21 @@
+class Common
+  attr_accessor :name
+  attr_accessor :eng
+  attr_accessor :price
+  attr_accessor :type
+  attr_accessor :anonymous
+  attr_accessor :tracking
+  attr_accessor :compensation
+  attr_accessor :note
+
+  def initialize(name:, eng:, price:, type:, anonymous:, tracking:, compensation:, note:)
+    @name = name
+    @eng = eng
+    @price = price
+    @type = type
+    @anonymous = anonymous
+    @tracking = tracking
+    @compensation = compensation
+    @note = note
+  end
+end
